requirements.txt
# FastAPI و Async Framework
fastapi>=0.110.0,<0.120.0
uvicorn==0.29.0
gunicorn==22.0.0

# Web Extensions
fastapi-limiter==0.1.6
python-multipart==0.0.9

# Database & Caching
redis>=5.0.0
aiocache==0.12.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5
python-dotenv==1.0.1

# AI & Processing
google-generativeai==0.4.1
requests==2.31.0
aiohttp==3.9.3
pydub==0.25.1

# Monitoring & Performance
prometheus-client==0.20.0
psutil==5.9.8
apscheduler==3.10.4  

# Data Handling
pydantic>=2.6.0,<3.0.0
typing-extensions==4.10.0
cachetools==5.3.3

# Audio & Multimedia
soundfile==0.12.1
ffmpeg-python==0.2.0

# Miscellaneous
aiofiles==24.1.0
pyyaml==6.0.1
starlette==0.37.2

# Development & Testing
pytest==8.1.1
coverage==7.4.4
httpx==0.27.0

# الأمان والتشفير
secure==0.3.0

itsdangerous==2.2.0
async-timeout==4.0.3
